+++ app/components/sidebars/drawer_layout.js
@@ -11,7 +11,6 @@ import {
     Easing,
     Keyboard,
     PanResponder,
-    Platform,
     StyleSheet,
     TouchableWithoutFeedback,
     View,
@@ -120,16 +119,15 @@ export default class DrawerLayout extends Component {
 
     componentDidMount() {
         Dimensions.addEventListener('change', this.handleDimensionsChange);
-        if (Platform.OS === 'ios') {
-            // on iOS force closing the drawers to prevent them for partially showing
-            // when channging the device orientation, probably caused by RN61
-            Animated.timing(this.openValue, {
-                toValue: 0,
-                duration: 10,
-                easing: Easing.out(Easing.cubic),
-                useNativeDriver: this.props.useNativeAnimations,
-            }).start();
-        }
+
+        // on iOS force closing the drawers to prevent them for partially showing
+        // when channging the device orientation, probably caused by RN61
+        Animated.timing(this.openValue, {
+            toValue: 0,
+            duration: 10,
+            easing: Easing.out(Easing.cubic),
+            useNativeDriver: this.props.useNativeAnimations,
+        }).start();
     }
 
     componentWillUnmount() {
@@ -199,7 +197,7 @@ export default class DrawerLayout extends Component {
         /* Drawer styles */
         let outputRange;
         // ios main sidebar sits mostly under the main screen, with slight move
-        const translateDistance = Platform.OS === 'ios' && drawerPosition === 'left' ?
+        const translateDistance = drawerPosition === 'left' ?
             Math.floor(drawerWidth * 0.2) : drawerWidth
 
         if (this.getDrawerPosition() === 'left') {
@@ -221,7 +219,7 @@ export default class DrawerLayout extends Component {
         // 1 - main | tablet
         // 2 - overlay
         // 3 - android main drawer | settings drawer
-        const drawerZIndex = drawerPosition === 'left' && Platform.OS === 'ios' ? 0 : 3
+        const drawerZIndex = drawerPosition === 'left' ? 0 : 3
 
         return (
             <React.Fragment>
@@ -269,7 +267,7 @@ export default class DrawerLayout extends Component {
         }
 
         const mainStyles = [styles.main]
-        if (drawerPosition === 'left' && Platform.OS === 'ios') {
+        if (drawerPosition === 'left') {
             /* Drawer styles */
             let outputRange;
             
@@ -417,10 +415,7 @@ export default class DrawerLayout extends Component {
                     return true;
                 }
             } else {
-                const filter = Platform.select({
-                    ios: moveX > 0 && dx > 35,
-                    android: moveX <= 35 && dx > 0
-                });
+                const filter = moveX > 0 && dx > 35;
                 if (filter) {
                     this._isClosing = false;
                     return true;
@@ -454,7 +449,7 @@ export default class DrawerLayout extends Component {
     };
 
     _panResponderMove = (e: EventType, { moveX, dx }: PanResponderEventType) => {
-        const useDx = Platform.OS === 'ios' && this.getDrawerPosition() === 'left' && !this._isClosing;
+        const useDx = this.getDrawerPosition() === 'left' && !this._isClosing;
         let openValue = this._getOpenValueForX(useDx ? dx : moveX);
 
         if (this._isClosing) {
