+++ app/components/sidebars/main/index.js
@@ -10,11 +10,11 @@ import {getTheme} from 'mattermost-redux/selectors/entities/preferences';
 import {getCurrentTeamId, getMyTeamsCount} from 'mattermost-redux/selectors/entities/teams';
 import {getCurrentUser} from 'mattermost-redux/selectors/entities/users';
 
-import {setChannelDisplayName, setChannelLoading, handleSelectChannel} from 'app/actions/views/channel';
+import {setChannelDisplayName, handleSelectChannel} from 'app/actions/views/channel';
 import {makeDirectChannel} from 'app/actions/views/more_dms';
 import telemetry from 'app/telemetry';
 
-import MainSidebar from './main_sidebar.js';
+import MainSidebar from './main_sidebar';
 
 export function logChannelSwitch(channelId, currentChannelId) {
     return (dispatch, getState) => {
@@ -54,7 +54,6 @@ function mapDispatchToProps(dispatch) {
             logChannelSwitch,
             makeDirectChannel,
             setChannelDisplayName,
-            setChannelLoading,
             handleSelectChannel,
         }, dispatch),
     };
