+++ app/screens/channel/channel.ios.js
@@ -2,19 +2,22 @@
 // See LICENSE.txt for license information.
 
 import React from 'react';
-import {Dimensions, View} from 'react-native';
+import {View} from 'react-native';
 import {KeyboardTrackingView} from 'react-native-keyboard-tracking-view';
 
 import Autocomplete, {AUTOCOMPLETE_MAX_HEIGHT} from 'app/components/autocomplete';
+import InteractiveDialogController from 'app/components/interactive_dialog_controller';
+import MainSidebar from 'app/components/sidebars/main';
 import NetworkIndicator from 'app/components/network_indicator';
 import PostTextbox from 'app/components/post_textbox';
 import SafeAreaView from 'app/components/safe_area_view';
+import SettingsSidebar from 'app/components/sidebars/settings';
 import StatusBar from 'app/components/status_bar';
-import {DeviceTypes} from 'app/constants';
+import {makeStyleSheetFromTheme} from 'app/utils/theme';
 
 import LocalConfig from 'assets/config';
 
-import ChannelBase, {ClientUpgradeListener, style} from './channel_base';
+import ChannelBase, {ClientUpgradeListener} from './channel_base';
 import ChannelNavBar from './channel_nav_bar';
 import ChannelPostList from './channel_post_list';
 
@@ -23,23 +26,47 @@ const CHANNEL_POST_TEXTBOX_CURSOR_CHANGE = 'onChannelTextBoxCursorChange';
 const CHANNEL_POST_TEXTBOX_VALUE_CHANGE = 'onChannelTextBoxValueChange';
 
 export default class ChannelIOS extends ChannelBase {
-    render() {
-        const {height} = Dimensions.get('window');
-        const {currentChannelId} = this.props;
+    mainSidebarRef = (ref) => {
+        if (ref) {
+            this.mainSidebar = ref;
+        }
+    };
 
-        const channelLoaderStyle = [style.channelLoader, {height}];
-        if ((DeviceTypes.IS_IPHONE_WITH_INSETS || DeviceTypes.IS_TABLET)) {
-            channelLoaderStyle.push(style.iOSHomeIndicator);
+    settingsSidebarRef = (ref) => {
+        if (ref) {
+            this.settingsSidebar = ref;
+        }
+    };
+
+    openMainSidebar = () => {
+        if (this.mainSidebar) {
+            this.mainSidebar.open();
+        }
+    };
+
+    openSettingsSidebar = () => {
+        if (this.settingsSidebar) {
+            this.settingsSidebar.open();
+        }
+    };
+
+    render() {
+        const {currentChannelId, theme} = this.props;
+
+        const channelLoadingOrFailed = this.renderLoadingOrFailedChannel();
+        if (channelLoadingOrFailed) {
+            return channelLoadingOrFailed;
         }
 
+        const style = getStyle(theme);
         const drawerContent = (
-            <React.Fragment>
+            <>
                 <SafeAreaView>
                     <StatusBar/>
                     <NetworkIndicator/>
                     <ChannelNavBar
-                        openChannelDrawer={this.openChannelSidebar}
-                        openSettingsDrawer={this.openSettingsSidebar}
+                        openMainSidebar={this.openMainSidebar}
+                        openSettingsSidebar={this.openSettingsSidebar}
                         onPress={this.goToChannelInfo}
                     />
                     <ChannelPostList
@@ -67,9 +94,30 @@ export default class ChannelIOS extends ChannelBase {
                         screenId={this.props.componentId}
                     />
                 </KeyboardTrackingView>
-            </React.Fragment>
+            </>
         );
 
-        return this.renderChannel(drawerContent);
+        return (
+            <MainSidebar ref={this.mainSidebarRef}>
+                <SettingsSidebar ref={this.settingsSidebarRef}>
+                    <View style={style.backdrop}>
+                        {drawerContent}
+                    </View>
+                </SettingsSidebar>
+                <InteractiveDialogController
+                    theme={theme}
+                />
+            </MainSidebar>
+        );
     }
 }
+
+export const getStyle = makeStyleSheetFromTheme((theme) => ({
+    backdrop: {
+        flex: 1,
+        backgroundColor: theme.centerChannelBg,
+    },
+    flex: {
+        flex: 1,
+    },
+}));
