+++ app/screens/channel/channel_nav_bar/channel_drawer_button/channel_drawer_button.test.js
@@ -39,7 +39,7 @@ jest.mock('react-native-notifications', () => {
 
 describe('ChannelDrawerButton', () => {
     const baseProps = {
-        openDrawer: jest.fn(),
+        openSidebar: jest.fn(),
         badgeCount: 0,
         theme: Preferences.THEMES.default,
         visible: false,
