+++ app/screens/channel/channel_nav_bar/channel_nav_bar.js
@@ -29,8 +29,8 @@ const {
 export default class ChannelNavBar extends PureComponent {
     static propTypes = {
         isLandscape: PropTypes.bool.isRequired,
-        openChannelDrawer: PropTypes.func.isRequired,
-        openSettingsDrawer: PropTypes.func.isRequired,
+        openMainSidebar: PropTypes.func.isRequired,
+        openSettingsSidebar: PropTypes.func.isRequired,
         onPress: PropTypes.func.isRequired,
         theme: PropTypes.object.isRequired,
     };
@@ -72,7 +72,7 @@ export default class ChannelNavBar extends PureComponent {
 
     render() {
         const {isLandscape, onPress, theme} = this.props;
-        const {openChannelDrawer, openSettingsDrawer} = this.props;
+        const {openMainSidebar, openSettingsSidebar} = this.props;
         const style = getStyleFromTheme(theme);
 
         let height;
@@ -107,7 +107,7 @@ export default class ChannelNavBar extends PureComponent {
         return (
             <View style={[style.header, padding(isLandscape), {height}]}>
                 <ChannelDrawerButton
-                    openDrawer={openChannelDrawer}
+                    openSidebar={openMainSidebar}
                     visible={drawerButtonVisible}
                 />
                 <ChannelTitle
@@ -117,7 +117,7 @@ export default class ChannelNavBar extends PureComponent {
                 <ChannelSearchButton
                     theme={theme}
                 />
-                <SettingDrawerButton openDrawer={openSettingsDrawer}/>
+                <SettingDrawerButton openSidebar={openSettingsSidebar}/>
             </View>
         );
     }
