+++ app/screens/channel/channel_post_list/channel_post_list.js
@@ -11,12 +11,10 @@ import {
 
 import {getLastPostIndex} from 'mattermost-redux/utils/post_list';
 import EventEmitter from 'mattermost-redux/utils/event_emitter';
-import {WebsocketEvents} from 'mattermost-redux/constants';
 
 import AnnouncementBanner from 'app/components/announcement_banner';
 import PostList from 'app/components/post_list';
 import RetryBarIndicator from 'app/components/retry_bar_indicator';
-import {ViewTypes} from 'app/constants';
 import tracker from 'app/utils/time_tracker';
 import {changeOpacity, makeStyleSheetFromTheme} from 'app/utils/theme';
 import telemetry from 'app/telemetry';
@@ -31,7 +29,6 @@ export default class ChannelPostList extends PureComponent {
             loadPostsIfNecessaryWithRetry: PropTypes.func.isRequired,
             loadThreadIfNecessary: PropTypes.func.isRequired,
             increasePostVisibility: PropTypes.func.isRequired,
-            increasePostVisibilityByOne: PropTypes.func.isRequired,
             selectPost: PropTypes.func.isRequired,
             recordLoadTime: PropTypes.func.isRequired,
             refreshChannelWithRetry: PropTypes.func.isRequired,
@@ -43,7 +40,6 @@ export default class ChannelPostList extends PureComponent {
         lastViewedAt: PropTypes.number,
         loadMorePostsVisible: PropTypes.bool.isRequired,
         postIds: PropTypes.array,
-        postVisibility: PropTypes.number,
         refreshing: PropTypes.bool.isRequired,
         theme: PropTypes.object.isRequired,
         updateNativeScrollView: PropTypes.func,
@@ -51,15 +47,10 @@ export default class ChannelPostList extends PureComponent {
 
     static defaultProps = {
         postIds: [],
-        postVisibility: ViewTypes.POST_VISIBILITY_CHUNK_SIZE,
     };
 
     constructor(props) {
         super(props);
-
-        this.state = {
-            visiblePostIds: this.getVisiblePostIds(props),
-        };
 
         this.contentHeight = 0;
 
@@ -69,29 +60,14 @@ export default class ChannelPostList extends PureComponent {
 
     componentDidMount() {
         EventEmitter.on('goToThread', this.goToThread);
-        EventEmitter.on(WebsocketEvents.INCREASE_POST_VISIBILITY_BY_ONE, this.increasePostVisibilityByOne);
-    }
-
-    componentWillReceiveProps(nextProps) {
-        const {postIds: nextPostIds} = nextProps;
-
-        let visiblePostIds = this.state.visiblePostIds;
-        if (nextPostIds !== this.props.postIds || nextProps.postVisibility !== this.props.postVisibility) {
-            visiblePostIds = this.getVisiblePostIds(nextProps);
-        }
-
-        if (this.props.channelId !== nextProps.channelId) {
-            this.isLoadingMoreTop = false;
-        }
-
-        if (this.state.visiblePostIds !== visiblePostIds) {
-            this.setState({visiblePostIds});
-        }
     }
 
     componentDidUpdate(prevProps) {
-        if (prevProps.channelId !== this.props.channelId && tracker.channelSwitch) {
-            this.props.actions.recordLoadTime('Switch Channel', 'channelSwitch');
+        if (this.props.channelId !== prevProps.channelId) {
+            this.isLoadingMoreTop = false;
+            if (tracker.channelSwitch) {
+                this.props.actions.recordLoadTime('Switch Channel', 'channelSwitch');
+            }
         }
 
         if (!prevProps.postIds?.length && this.props.postIds?.length > 0 && this.props.updateNativeScrollView) {
@@ -102,12 +78,7 @@ export default class ChannelPostList extends PureComponent {
 
     componentWillUnmount() {
         EventEmitter.off('goToThread', this.goToThread);
-        EventEmitter.off(WebsocketEvents.INCREASE_POST_VISIBILITY_BY_ONE, this.increasePostVisibilityByOne);
     }
-
-    getVisiblePostIds = (props) => {
-        return props.postIds ? props.postIds.slice(0, props.postVisibility) : [];
-    };
 
     goToThread = (post) => {
         telemetry.start(['post_list:thread']);
@@ -130,18 +101,13 @@ export default class ChannelPostList extends PureComponent {
         });
     };
 
-    increasePostVisibilityByOne = () => {
-        const {actions, channelId} = this.props;
-        actions.increasePostVisibilityByOne(channelId);
-    }
-
     loadMorePostsTop = () => {
-        const {actions, channelId} = this.props;
+        const {actions, channelId, postIds} = this.props;
         if (!this.isLoadingMoreTop) {
             this.isLoadingMoreTop = true;
             actions.increasePostVisibility(
                 channelId,
-                this.state.visiblePostIds[this.state.visiblePostIds.length - 1],
+                postIds[postIds.length - 1],
             ).then((hasMore) => {
                 this.isLoadingMoreTop = !hasMore;
             });
@@ -190,15 +156,14 @@ export default class ChannelPostList extends PureComponent {
             channelRefreshingFailed,
             currentUserId,
             lastViewedAt,
-            loadMorePostsVisible,
+            postIds,
             refreshing,
             theme,
         } = this.props;
 
-        const {visiblePostIds} = this.state;
         let component;
 
-        if (visiblePostIds.length === 0 && channelRefreshingFailed) {
+        if (postIds.length === 0 && channelRefreshingFailed) {
             const FailedNetworkAction = require('app/components/failed_network_action').default;
 
             component = (
@@ -210,9 +175,9 @@ export default class ChannelPostList extends PureComponent {
         } else {
             component = (
                 <PostList
-                    postIds={visiblePostIds}
-                    lastPostIndex={Platform.OS === 'android' ? getLastPostIndex(visiblePostIds) : -1}
-                    extraData={loadMorePostsVisible}
+                    postIds={postIds}
+                    lastPostIndex={Platform.OS === 'android' ? getLastPostIndex(postIds) : -1}
+                    extraData={postIds.length === 0}
                     onLoadMoreUp={this.loadMorePostsTop}
                     onPostPress={this.goToThread}
                     onRefresh={actions.setChannelRefreshing}
